test: InputService Test 코드 Trial으로 이동
